refactor(backend): move pre-turn hooks logic to dedicated module
- Create app/core/agent_hooks.py to house pre-turn hook logic
- Remove build_pre_turn_hooks from app/core/agent_tools.py
- Update imports in api/chat.py and telegram/bot.py
- Refactor typing indicator refresh interval calculation
